feat(ISSUE-27): Adding some simple file cache tests
